backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "carspark-backend"
version = "1.0.0"
description = "Django backend for CarSpark project"
authors = [{name = "CarSpark Team"}]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "Django>=4.2.0",
    "djangorestframework>=3.14.0",
    "django-filter>=23.0",
    "django-cors-headers>=4.0.0",
    "python-decouple>=3.8",
    "boto3>=1.26.0",
    "django-storages>=1.13.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "pytest>=7.0.0",
    "pytest-django>=4.5.0",
]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_django = "django"
known_first_party = ["cars", "carspark_backend"]
sections = ["FUTURE", "STDLIB", "DJANGO", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "carspark_backend.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = "-v --tb=short"
